HF_Streamlit/requirements.txt
langchain
python-dotenv
langchain-community
langchain-core
bs4
faiss-cpu
streamlit
pypdf